+++ pkg/topic/controller/datastore/tables.go
@@ -39,7 +39,7 @@ func (d *TopicDatastore) createTopicTables(ctx context.Context, tx pgx.Tx, id in
 			-- claim fence assumes ids are issued in INSERT order, and a cached
 			-- sequence hands out out-of-order id blocks
 
-			schema_version BIGINT NOT NULL,               -- the payload's version, from the producing Message type
+			schema_version INTEGER NOT NULL,              -- the payload's version, from the producing Message type
 			routing_key TEXT,
 			message_key TEXT,
 			compaction_rank BIGINT,                       -- NULL = this message never opted into compaction
@@ -223,7 +223,7 @@ func (d *TopicDatastore) createTopicTables(ctx context.Context, tx pgx.Tx, id in
 		CREATE TABLE IF NOT EXISTS %[1]s.%[2]s (
 			compaction_key  TEXT   NOT NULL PRIMARY KEY,
 			message_id      BIGINT,                    -- NULL while the key has a lockable row but no winning message
-			schema_version  BIGINT,                    -- the winner's payload version; compared before rank
+			schema_version  INTEGER,                   -- the winner's payload version; compared before rank
 			compaction_rank BIGINT,                    -- the winner's rank
 			created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 			updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
